Djreact/templates/DReact/Boilerplate/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
django-allauth==0.55.0
dj-rest-auth==2.1.9
social-auth-app-django==4.0.0
django-filter==23.1
psycopg2==2.9.5
stripe==5.0.1
python-decouple==3.8
redis==4.3.4
jazzmin==3.1.0

